% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profiler_ORA_per_group.R
\name{cluster_profiler_ORA_per_group}
\alias{cluster_profiler_ORA_per_group}
\title{Perform enrichment analysis on WGCNA gene lists using the latest version of clusterProfiler}
\usage{
cluster_profiler_ORA_per_group(
  df,
  gene_id_cols = "hgnc_symbol",
  group_column = "gene_groups",
  organism = "Homo sapiens",
  groups = NULL,
  ids_type = c("hgnc_symbol", "ensembl_gene_id"),
  export_to_xlsx = FALSE,
  collections = c("H", "C2;CP", "C5;GO:BP"),
  pvalueCutoff = 0.05,
  export_fdr_thres = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 1,
  useBackground = T
)
}
\arguments{
\item{df}{Table containing gene names and the group they belong}

\item{gene_id_cols}{The name of the column containing gene ids (HGNC_Symbols)}

\item{group_column}{The name of the column defining the gene groups}

\item{organism}{The organism to search for: Homo sapiens or Mus musculus}

\item{groups}{A vector containing the modules to be tested. If this is given, any other module parameter is overriden.}

\item{export_to_xlsx}{TRUE/FALSE. Whether to export results for each module to an excel file.}

\item{pvalueCutoff}{ClusterProfiler's adjusted pvalue cutoff on enrichment tests to report}

\item{export_fdr_thres}{Used when export to xlsx is TRUE to filter the final list of geneset results to be exported. By default, it equals the pvalueCutoff.}

\item{pAdjustMethod}{ClusterProfiler's: one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{minGSSize}{ClusterProfiler's minimal size of genes annotated for testing}

\item{maxGSSize}{ClusterProfiler's maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.}

\item{genesets}{The collections of MSigDB R package to check enrichment of in the format of Category;Subcategory.}

\item{fdr_threshold}{The FDR threshold for results exported in Excel file}
}
\value{
A list with enrichment results as elements.
}
\description{
Perform enrichment analysis on WGCNA gene lists using the latest version of clusterProfiler
}
\details{
Dependencies are:
library(clusterProfiler)
library(openxlsx)
library(dplyr)
library(stringr)
library(msigdbr)

date 02/05/2024
}
