% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap2_cc_ss_plot.R
\name{cmap2_cc_ss_plot}
\alias{cmap2_cc_ss_plot}
\title{Plot Signature strength over replicate correlation based on a Cmap2 results table}
\usage{
cmap2_cc_ss_plot(
  cmap_melted,
  n_top_highlight = NULL,
  cell_types = NULL,
  ids_to_highlight = NULL,
  max.overlaps = 40,
  min.segment.length = unit(0.001, "mm"),
  label_alpha = 0.8,
  label_size = 3
)
}
\arguments{
\item{cmap_melted}{A melted data.frame containing Connectivity Map 2 results}

\item{n_top_highlight}{Number of top results to label}

\item{cell_types}{Cell types to keep}

\item{ids_to_highlight}{Particular perturbation ids to label. Overrides n_top_highlight if used.}

\item{max.overlaps}{geom_label_repel max overlaps parameter}

\item{min.segment.length}{geom_label_repel minimum segment length parameter}

\item{label_alpha}{Transparency parameter for labels}

\item{label_size}{Size parameter for labels}
}
\value{
A ggplot
}
\description{
Dependencies are:
library(cmapR)
library(tidyverse)
library(ggrepel)
}
\details{
date 22/04/2024
}
