% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap2_heatmap_create.R
\name{cmap2_heatmap_create}
\alias{cmap2_heatmap_create}
\title{Plot a heatmap based on a Cmap2 results table}
\usage{
cmap2_heatmap_create(
  cmap_melted,
  n_top = 20,
  direction = c("both", "normal", "reverse"),
  CS = NULL,
  brewer_palette_qual = "Paired",
  row_label_size = 7,
  col_label_size = 9,
  show_row_dend = T,
  heatmap_title = "NULL",
  legend_title = "Normalized\\nConnectivity\\nScore",
  legend_breaks = c(-3, -1.5, 0, 1.5, 3),
  label_column = c("pert_iname", "pert_id"),
  row_annot_cols = c("moa", "pert_idose", "pert_itime", "pert_type", "target_name",
    "tas", "cc_q75", "fdr_q_nlog10", "raw_cs", "cell_iname"),
  border_color = "black",
  border_linetype = 1,
  border_linewidth = 1,
  cell_border_color = "grey90",
  cell_border_linewidth = 1
)
}
\arguments{
\item{cmap_melted}{A melted data.frame containing Connectivity Map 2 results}

\item{n_top}{Number of top results to keep}

\item{direction}{What direction of signatures should be kept. Normal keeps mimicking signatures, reverse keeps reversing signatures, both keeps both.}

\item{CS}{Heatmap color scale function}

\item{brewer_palette_qual}{Qualitative brewer palette for mapping qualitative variables. Check out brewer_pal_info for more info.}

\item{row_label_size}{The size of row labels}

\item{col_label_size}{The size of column labels}

\item{show_row_dend}{Whether to show row dendrogram}

\item{heatmap_title}{Heatmap title}

\item{legend_title}{How to name the heatmap color scale}

\item{legend_breaks}{Which legend breaks to display}

\item{label_column}{Which column to use for row labels}

\item{row_annot_cols}{Which columns to keep as row annotation}

\item{border_color}{Color of the heatmap border}

\item{border_linetype}{Linetype of the heatmap border}

\item{border_linewidth}{Linewidth of the heatmap border}

\item{cell_border_color}{Color of the cell border}

\item{cell_border_linewidth}{Linewidth of the cell border}
}
\value{
A heatmap
}
\description{
Dependencies are:
library(cmapR)
library(dplyr)
library(tibble)
library(tidyr)
library(ComplexHeatmap)
library(circlize)
library(RColorBrewer)
}
\details{
date 22/04/2024
}
