% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap2_matrices_read.R
\name{cmap2_matrices_read}
\alias{cmap2_matrices_read}
\title{Parse Cmap2 results into a table}
\usage{
cmap2_matrices_read(clue_results_dir, keep_only_morpheus = FALSE)
}
\arguments{
\item{clue_results_dir}{Directory path of the downloaded CLUE Cmap2.0 results}

\item{keep_only_morpheus}{Whether to keep only rows corresponding to the online CLUE Morpheus visualization (essentially the same table as the query_result.gct in the ARFS folder, with the addition of the fdr column)}
}
\value{
A dataframe
}
\description{
Dependencies are:
library(cmapR)
library(dplyr)
library(tibble)
library(tidyr)
}
\details{
date 19/04/2024
}
