% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_htseq_files.R
\name{combine_htseq_files}
\alias{combine_htseq_files}
\title{Create a count table}
\usage{
combine_htseq_files(indir, pattern = "*.txt", remove_last_5 = TRUE)
}
\arguments{
\item{pattern}{The pattern of file extensions to check. Defaults to '.txt'}

\item{remove_last_5}{Whether to remove the last 5 rows containing ambiguous (etc.) counts or not}

\item{path}{Folder containing HTSeq output files}
}
\value{
count table
}
\description{
Creates a count table using multiple HTSeq output files as input
}
\details{
date 08/03/2023
}
