% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_tpm.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Convert raw counts to tpm counts}
\usage{
counts_to_tpm(
  df,
  identifier_type = c("hgnc_symbol", "ensembl_gene_id", "mgi_symbol"),
  ensembl_version = NULL,
  gene_lengths = NULL,
  organism = c("human", "mouse")
)
}
\arguments{
\item{df}{Input data frame with raw counts and gene ids as rownames}

\item{identifier_type}{One of hgnc_symbol, ensembl_gene_id  of mgi_symbol}

\item{ensembl_version}{The version of ensembl to be used. If none is provided the function uses the default (latest) version}

\item{gene_lengths}{A vector of the same length as the rows of the df containing the length of each gene}

\item{organism}{The organism to search for: human or mouse}
}
\value{
A dataframe with TPM expression values
}
\description{
This function takes a table with raw expression counts and converts them to TPM
This function does not take into account effective gene length.
Dependencies are:
library(biomaRt)
}
\details{
date 30/04/2023
}
