% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_tpm_94.R
\name{counts_to_tpm_94}
\alias{counts_to_tpm_94}
\title{Convert raw counts to tpm counts}
\usage{
counts_to_tpm_94(
  df,
  annot,
  check.names = FALSE,
  gene_col_id = "Gene stable ID version"
)
}
\arguments{
\item{df}{Input data frame with raw counts and gene ids as rownames}

\item{annot}{A dataframe with annotation containing Gene start (bp) or Gene end (bp) columns to determine gene length. This should also contain a column with the gene ids (as found on the df input) to match the dfs.}

\item{check.names}{Whether to use 'Gene.start.(bp)' or 'Gene start (bp)'}

\item{gene_col_id}{The name of the column that corresponds to the gene ids as they are found on the rownames of the df input.}
}
\value{
A dataframe with TPM expression values
}
\description{
This function takes a table with raw expression counts and converts them to TPM
This function does not take into account effective gene length.
Dependencies are:
}
\details{
date 09/04/2024
}
