% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rnk.R
\name{create_rnk}
\alias{create_rnk}
\title{Dependencies are:
none}
\usage{
create_rnk(
  lrt_table,
  logFC_col_name = "logFC",
  sig_measure_col_name = "fdr",
  remove_version = FALSE,
  gene_id_column = NULL,
  output_table_filename = NULL,
  return_data = FALSE
)
}
\arguments{
\item{logFC_col_name}{}

\item{sig_measure_col_name}{}

\item{gene_id_column}{}

\item{return_data}{}
}
\value{
Creates an rnk file in the directory (current or specified)
}
\description{
date 21/07/2023
}
