% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_eset.R
\name{dedup_eset}
\alias{dedup_eset}
\title{Deduplicate probes of Expression Set Microarray object}
\usage{
dedup_eset(eset, fData_gene_symbol_column_name = "Gene Symbol")
}
\arguments{
\item{eset}{Input ExpressionSet object}

\item{fData_gene_symbol_column_name}{The name of the column containing the gene symbols in feature Data of the ExpressionSet object}
}
\value{
An ExpressionSet object with 1 probe per gene
}
\description{
This function takes an Expression Set object, finds probes that correspond to the same gene 
according to fData, and deduplicates.
}
\details{
Deduplication is performed based on higher expression and if that is a tie then 
it chooses higher variation.
If only 1 probe has a mean expression over 1, removes all others. 
If 2 or more probes have expression over 1, or 2 or all probes have expression 
lower than 1, then it keeps the one with the highest variance.

Dependencies are:
library(Biobase)

date 10/07/2023
}
