% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSG_manipulator.R
\name{ENSG_manipulator_shiny}
\alias{ENSG_manipulator_shiny}
\title{Run an interactive session to remove ENSG ID version number}
\usage{
ENSG_manipulator_shiny()
}
\value{
Opens a shiny session
}
\description{
Dependencies are:
library(shiny)
library(biomaRt)
}
\details{
date 24/02/2023

to specify repositories for deployment use options(repos = BiocManager::repositories())
}
