% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensg_to_hgnc.R
\name{ensg_to_hgnc}
\alias{ensg_to_hgnc}
\title{Convert ENSG_ID rownames of a table to HGNC_ID IDs}
\usage{
ensg_to_hgnc(
  table,
  ensembl_version = 0,
  organism = "hsapiens",
  version = FALSE
)
}
\arguments{
\item{table}{Input table with ENSG IDs as rownames}

\item{ensembl_version}{The version of ensembl to be used. If none is provided the function uses the default (latest) version}

\item{organism}{The organism to search for: hsapiens or mmusculus}

\item{version}{TRUE/FALSE. Whether the identifiers of the original dataframe have an ensembl gene id version number}
}
\value{
A dataframe with deduplicated HGNC IDs as rownames
}
\description{
This function takes a table with Ensembl Gene IDs, converts them to HGNC symbols, 
leaves out all non-converted IDs (Blanks) and deduplicates
}
\details{
Deduplication is performed based on higher expression and if that is a tie then 
it chooses higher variation.
If only 1 gene has a mean expression over 1, removes all others. 
If 2 or more genes have expression over 1, or 2 or all genes have expression 
lower than 1, then keeps the one with the highest variance.

Dependencies are:
library(biomaRt)

date 04/07/2022
}
