% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_counts.R
\name{generate_counts}
\alias{generate_counts}
\title{Read counts from multiple files}
\usage{
generate_counts(
  count_files_dir,
  file_ext_to_remove,
  remove_last_rows = FALSE,
  rows_to_remove = 4,
  sep = "\\t"
)
}
\arguments{
\item{dir}{Directory of count files}
}
\value{
Data Frame
}
\description{
date 06/01/2023
}
