% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_vms.R
\name{get_gene_vms}
\alias{get_gene_vms}
\title{Get gene Variance, Mean and Sum.}
\usage{
get_gene_vms(df, genes)
}
\arguments{
\item{df}{Data frame containing gene expression with gene names as rownames.}

\item{genes}{The list of genes to gather Variance, Mean and Sum.}
}
\value{
A Data frame containing gene Variance, Mean and Sum.
}
\description{
George Sentis
}
\details{
date 31/01/2024
}
