% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_sen.R
\name{heatmap_sen}
\alias{heatmap_sen}
\title{Dependencies are:
library(ComplexHeatmap)
library(gridtext)}
\usage{
heatmap_sen(
  norm_expr,
  lrt_table = NULL,
  col_fun = NULL,
  z_scaling = TRUE,
  samples = NULL,
  ntop = NULL,
  order_factor = "logFC",
  title = NULL,
  legend_title = NULL,
  col_annotation = NULL,
  annotcolors = NULL,
  col_name_size = 8,
  row_name_size = 8,
  convert_to_hgnc = FALSE,
  ensembl_version = 0,
  identifier_versions = FALSE,
  conversion_organism = "hsapiens",
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  annotation_legend = NULL,
  title_render = TRUE,
  box_fill = "#EEEEEE"
)
}
\arguments{
\item{norm_expr}{}

\item{col_fun}{}

\item{z_scaling}{}

\item{ntop}{}

\item{legend_title}{}

\item{col_annotation}{}

\item{col_name_size}{}

\item{box_fill}{}
}
\value{
A ComplexHeatmap
}
\description{
date 21/07/2023
}
