% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_genes.R
\name{isolate_genes}
\alias{isolate_genes}
\title{Isolate a subset of genes from a gene expression table}
\usage{
isolate_genes(object, gene_list, gene_column = "rownames")
}
\arguments{
\item{object}{Gene expression table}

\item{gene_list}{Vector with gene names to be isolated}

\item{gene_column}{Column in the gene expression table that contains the gene names}
}
\value{
Same class as object but subsetted
}
\description{
date 08/02/2023
}
