% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_sen_vis.R
\name{pca_sen_vis}
\alias{pca_sen_vis}
\title{Create a pca plot from an existing pca_object}
\usage{
pca_sen_vis(
  pca_object,
  color_factor,
  color_factor_name = "Color",
  colors_list = NULL,
  shape_vector = NULL,
  stroke = 2,
  return_data = FALSE,
  title = "PCA",
  PCx = 1,
  PCy = 2,
  point_alpha = 0.7,
  point_size = 7,
  add_ellipse = FALSE,
  save_pdf = "FALSE",
  pdf_height = 10,
  pdf_width = 10,
  pdf_file = "PCA_sen_output.pdf"
)
}
\arguments{
\item{pca_object}{PCA list created using pca_sen with return_data=TRUE}

\item{color_factor}{A vector of a parameter to color the samples by}

\item{color_factor_name}{A title for the legend of the color parameter}

\item{colors_list}{A list of colors if you wish to manually specify group colors}

\item{title}{Title of the plot}

\item{PCx}{The x-axis Principal Component}

\item{PCy}{The y-axis Principal Component}

\item{point_alpha}{Set the transparency of points}

\item{point_size}{Set the size of points}

\item{add_ellipse}{If you wish to add an ellipse around the points}

\item{save_pdf}{Whether to save the plot to a pdf file}

\item{pdf_height}{The height of the output pdf}

\item{pdf_width}{The width of the output pdf}

\item{pdf_file}{The name or path of the output pdf}
}
\value{
ggplot object
}
\description{
Creates a pca plot based on an existing pca object
}
\details{
Dependencies are:
library(ggplot2)
library(FactoMineR)


date 07/11/2022
}
