% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_expression.R
\name{plot_genes_expression}
\alias{plot_genes_expression}
\title{Plot histograms of expression of specific genes}
\usage{
plot_genes_expression(expression, genes)
}
\arguments{
\item{expression}{Object with expression values. The row names of the object must be the gene identifiers.}

\item{genes}{The genes that will be plotted. Identifiers must match some row names of the expression object.}
}
\value{
Data Frame
}
\description{
Dependencies are:
library(ggplot2)
library(tidyverse)
}
\details{
date 10/02/2023
}
