% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gprofiler_multi.R
\name{read_gprofiler_multi}
\alias{read_gprofiler_multi}
\title{Read multiple gProfiler output csv files}
\usage{
read_gprofiler_multi(
  enrich_folder,
  no_of_top_results = 10000,
  pvalue_thres = 0.05,
  file_labels = NULL
)
}
\arguments{
\item{enrich_folder}{The directory containing the gprofiler files to be read.}

\item{no_of_top_results}{If you need filtering, how many top results to keep from each file. 10000 usually results in all terms being collected.}

\item{pvalue_thres}{The p value significance threshold for the enriched terms}

\item{file_labels}{The labels of the files. This is going to be converted into a column of the final table. This will make it easier to identify which file the enriched term originated from.}
}
\value{
A dataframe
}
\description{
Dependencies are:
none
}
\details{
date 12/07/2023
}
