% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_version.R
\name{remove_version}
\alias{remove_version}
\title{Isolate a subset of genes from a gene expression table}
\usage{
remove_version(ensg_ids)
}
\arguments{
\item{ensg_ids}{Vector of Ensembl IDs with version number}
}
\value{
Vector of Ensembl IDs without version number
}
\description{
date 08/02/2023
}
