% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHeatmap_sen.R
\name{shinyHeatmap_sen}
\alias{shinyHeatmap_sen}
\title{Run an interactive ComplexHeatmap session}
\usage{
shinyHeatmap_sen()
}
\value{
Opens a shiny session
}
\description{
Dependencies are:
library(shiny)
library(ComplexHeatmap)
library(circlize)
library(dplyr)
library(openxlsx)
library(edgeR)
library(stringr)
}
\details{
date 14/02/2023
}
