% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPCA_sen.R
\name{shinyPCA_sen}
\alias{shinyPCA_sen}
\title{Run an interactive PCA plot session}
\usage{
shinyPCA_sen()
}
\value{
Opens a shiny session
}
\description{
Dependencies are:
library(shiny)
library(DT)
library(plotly)
library(openxlsx)
library(tidyverse)
library(sentisR)
library(FactoMineR)
library(edgeR)
}
\details{
date 10/12/2023
}
