% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_up_down.R
\name{top_up_down}
\alias{top_up_down}
\title{Choose a number of up and down regulated genes based on logFC from a certain comparison.}
\usage{
top_up_down(
  lrt_table,
  order_factor = "logFC",
  ntop = 100,
  ntop_up = NULL,
  ntop_down = NULL,
  gene_id_column = NULL
)
}
\arguments{
\item{order_factor}{}

\item{ntop}{}

\item{ntop_down}{}
}
\value{
A vector or data.frame(when both ntop_up and ntop_down are used)
}
\description{
Dependencies are:
library(dplyr)
}
\details{
date 21/07/2023
}
