% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_sen.R
\name{volcano_sen}
\alias{volcano_sen}
\title{Create a volcano plot}
\usage{
volcano_sen(
  table,
  title = "Cond_Up vs Cond_Down",
  significanceMeasure = c("fdr", "PValue"),
  thresSig = 0.05,
  thresLogFC = 0.58,
  colorUp = "darkgreen",
  colorDown = "steelblue",
  colorNeutral = "lightgrey",
  xlim = c(-10, 10),
  ylim = c(0, 15),
  breaks_x_default = FALSE,
  breaks_y_default = FALSE,
  genes_to_label = NULL,
  label_columns_names = NULL,
  ids = c("single", "mixed"),
  label_size = 4,
  label_alpha = 0.7,
  max_overlaps = 40
)
}
\arguments{
\item{table}{Input table with logFC and (fdr or PValue) columns}

\item{title}{Title of the plot}

\item{significanceMeasure}{the significance Measure to be used for statistical significance. Must match the corresponding column' name}

\item{thresSig}{The statistical Measure significance value}

\item{thresLogFC}{The Log Fold change significance threshold}

\item{colorUp}{The upregulated genes colors}

\item{colorDown}{The downregulated genes colors}

\item{colorNeutral}{The color of non-statistically significant or non-logFC significant genes}

\item{xlim}{The limits of the x-axis}

\item{ylim}{The limits of the y-axis}

\item{breaks_x_default}{If you wish to use the default x-axis breaks}

\item{breaks_y_default}{If you wish to use the default y-axis breaks}

\item{genes_to_label}{The list of genes to label}

\item{ids}{If you wish to have a mixed type of ids (i.e. Ensembl and HGNC symbols) or a single type}

\item{label_size}{The size of the gene labels}

\item{label_alpha}{The transparency of the gene labels}

\item{max_overlaps}{The number of geom_label_repel max.overlaps parameter}

\item{label_column_names}{The columns of the gene labels}
}
\value{
ggplot object
}
\description{
Creates a volcano plot based on a gene list with logFC and (fdr or PValue) columns
Works best with edgeR results.
}
\details{
Dependencies are:
library(ggplot2)
library(dplyr)
library(ggrepel)


date 24/10/2022
}
