% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_hub_gene_plot.R
\name{wgcna_hub_gene_plot}
\alias{wgcna_hub_gene_plot}
\title{Plot Trait Gene Significance over Module Membership using the output of WGCNA}
\usage{
wgcna_hub_gene_plot(
  wgcna_res,
  module,
  trait,
  module_column = "moduleColor",
  MM.threshold_low = 0.7,
  MM.threshold_high = 1,
  GS.threshold_low = -0.5,
  GS.threshold_high = 0.5,
  label_col = "geneSymbol",
  max.overlaps = 10,
  min.segment.length = 0.01,
  add_labels = TRUE,
  color_by_module = TRUE
)
}
\arguments{
\item{wgcna_res}{Input wgcna results data.frame}

\item{module}{The name of the module you wish to plot. It has to match the MM.module column in your wgnca_res object}

\item{trait}{The name of the trait you wish to plot. It has to match the GS.trait column in your wgnca_res object}

\item{module_column}{The name of the column containing the information of which gene belongs to which column}

\item{MM.threshold_low}{The lower threshold of Module membership to limit the X axis}

\item{MM.threshold_high}{The upper threshold of Module membership to limit the X axis}

\item{GS.threshold_low}{The lower threshold of Gene significance to limit the Y axis}

\item{GS.threshold_high}{The upper threshold of Gene significance to limit the Y axis}

\item{label_col}{The name of the column containing the labels you wish to annotate the plot with (i.e. Gene Symbols)}

\item{max.overlaps}{Parameter for geom_label_repel}

\item{min.segment.length}{Parameter for geom_label_repel}

\item{add_labels}{Whether to use geom_label_repel to add labels to the points}

\item{color_by_module}{Whether to color the points according to the module color}
}
\value{
A ggplot object
}
\description{
Dependencies are:
library(ggplot2)
library(ggrepel)
library(stringr)
}
\details{
date 11/07/2023
}
