% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_module_stringdb_network_create.R
\name{wgcna_module_stringdb_network_create}
\alias{wgcna_module_stringdb_network_create}
\title{Create a STRINGdb PPI network based on genes in selected WGCNA modules}
\usage{
wgcna_module_stringdb_network_create(
  wgcna_gene_info,
  moduleColors,
  module_color_col = "moduleColor",
  string_version = "12.0",
  species = 9606,
  score_threshold = 500,
  network_type = c("physical", "full"),
  geneNameCol = "geneSymbol"
)
}
\arguments{
\item{wgcna_gene_info}{Data frame containing the results of gene-module membership of WGCNA results}

\item{moduleColors}{Names of the modules to be used for the network}

\item{module_color_col}{Name of the column containing the gene module color}

\item{string_version}{Version of the STRINGdb to be queried}

\item{species}{Species to be queried - 9606 is H.sapiens}

\item{score_threshold}{Interaction score threshold}

\item{network_type}{Physical of full}

\item{geneNameCol}{Name of the column containing the gene names}
}
\value{
An igraph network object
}
\description{
Dependencies are:
library(tidyverse)
library(STRINGdb)
library(igraph)
}
\details{
date 23/04/2024
}
