% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_module_stringdb_network_plot.R
\name{wgcna_module_stringdb_network_plot}
\alias{wgcna_module_stringdb_network_plot}
\title{Create a STRINGdb PPI network based on genes in selected WGCNA modules}
\usage{
wgcna_module_stringdb_network_plot(
  net,
  layout = "igraph",
  algorithm = "kk",
  seed = 7,
  edge_alpha = 0.05,
  node_size_vec = NULL,
  node_shape = 21,
  node_stroke = 1,
  node_alpha = 0.7,
  node_text_size_vec = NULL,
  pdf_width = 14,
  pdf_height = 14,
  pdf_filename = NULL,
  print_to_pdf = F
)
}
\arguments{
\item{net}{The network object to be plotted}

\item{layout}{layout package}

\item{algorithm}{igraph layout algorithm to be used}

\item{seed}{Seed to be used}

\item{edge_alpha}{Alpha value for edge transparency}

\item{node_size_vec}{A vector specifying node sizes. By default it is a function of their degree}

\item{node_shape}{Shape of the nodes}

\item{node_stroke}{Node circle line width}

\item{node_alpha}{Alpha value for node transparency}

\item{node_text_size_vec}{A vector specifying node text sizes. By default it is a function of their degree}

\item{pdf_width}{If print_to_pdf = T, specifies pdf width}

\item{pdf_height}{If print_to_pdf = T, specifies pdf height}

\item{pdf_filename}{If print_to_pdf = T, specifies pdf filename}

\item{print_to_pdf}{Whether to export to pdf file}
}
\value{
A ggplot object
}
\description{
Dependencies are:
library(tidyverse)
library(ggraph)
library(igraph)
}
\details{
date 23/04/2024
}
