% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_tpm.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Convert ENSG_ID rownames of a table to HGNC_ID IDs}
\usage{
counts_to_tpm(
  df,
  identifier_type = c("hgnc_symbol", "ensembl_gene_id", "mgi_symbol"),
  ensembl_version = NULL,
  gene_lengths = NULL,
  organism = c("human", "mouse")
)
}
\arguments{
\item{df}{Input data frame with ENSG IDs or HGNC symbols as rownames}

\item{ensembl_version}{The version of ensembl to be used. If none is provided the function uses the default (latest) version}

\item{organism}{The organism to search for: human or mouse}
}
\value{
A dataframe with TPM expression values
}
\description{
This function takes a table with Ensembl Gene IDs or HGNC symbolsn and converts the raw expression values to HGNC symbols TPM 
This function does not take into account effective gene length. Maybe you want to look more into this for scientific and publication purposes.
Dependencies are:
library(biomaRt)
}
\details{
date 30/04/2023
}
