% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_enrich_modules.R
\name{wgcna_enrich_modules}
\alias{wgcna_enrich_modules}
\title{Perform enrichment analysis on WGCNA gene lists using the latest version of gprofiler2}
\usage{
wgcna_enrich_modules(
  wgcna_gene_info,
  gene_id_cols = "EnsemblGeneID",
  module_color_column = "moduleColor",
  organism = "hsapiens",
  modules = NULL,
  module_Results = NULL,
  significant_only = TRUE,
  export_to_xlsx = FALSE
)
}
\arguments{
\item{wgcna_gene_info}{Table containing gene names and the module they belong}

\item{gene_id_cols}{The name of the column containing gene ids}

\item{organism}{The organism to search for: hsapiens or mmusculus}

\item{modules}{A vector containing the modules to be tested. If this is given, any other module parameter is overriden.}

\item{module_Results}{A data frame with module names as rownames containing columns with correlation p.values between module eigengenes and phenotypical traits.}

\item{significant_only}{TRUE/FALSE. Whether to check only modules that are significantly correlated with at least 1 phenotypical trait or check every module detected.}

\item{export_to_xlsx}{TRUE/FALSE. Whether to export results for each module to an excel file.}
}
\value{
A list with enrichment results as elements.
}
\description{
Perform enrichment analysis on WGCNA gene lists using the latest version of gprofiler2
}
\details{
Dependencies are:
library(gprofiler2)
library(openxlsx)

date 02/08/2023
}
